Feedback form validation: missing references
