+++ src/components/FeedbackForms/MissingReferences/index.ts
@@ -1,1 +1,1 @@
-export * from './ReferencesTable';
+export * from './MissingReferenceForm';
