+++ src/pages/feedback/missingreferences.tsx
@@ -1,18 +1,8 @@
-import { Flex, FormControl, FormLabel, HStack, Input, Button, Text, Box } from '@chakra-ui/react';
-import { feedbackItems, FeedbackLayout, SimpleLink } from '@components';
-import { ReferencesTable } from '@components/FeedbackForms/MissingReferences';
-import { useStore } from '@store';
+import { Text } from '@chakra-ui/react';
+import { feedbackItems, FeedbackLayout, MissingReferenceForm, SimpleLink } from '@components';
 import { NextPage } from 'next';
-import { ChangeEvent, useState } from 'react';
 
 const MissingReferences: NextPage = () => {
-  const username = useStore((state) => state.getUsername());
-  const [email, setEmail] = useState<string>(username ?? '');
-
-  const handleEmailChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setEmail(e.target.value);
-  };
-
   return (
     <FeedbackLayout title="Submit missing references for SciX Abstract Service">
       <Text my={2}>Please use this form to submit one or more citations currently missing from our databases.</Text>
@@ -28,27 +18,7 @@ const MissingReferences: NextPage = () => {
         </SimpleLink>
         for it first.
       </Text>
-      <Box as="form" my={2}>
-        <Flex direction="column" gap={4}>
-          <HStack gap={2}>
-            <FormControl isRequired>
-              <FormLabel>Name</FormLabel>
-              <Input></Input>
-            </FormControl>
-            <FormControl isRequired>
-              <FormLabel>Email</FormLabel>
-              <Input type="email" value={email} onChange={handleEmailChange}></Input>
-            </FormControl>
-          </HStack>
-          <ReferencesTable />
-          <HStack mt={2}>
-            <Button type="submit">Submit</Button>
-            <Button type="reset" variant="outline">
-              Reset
-            </Button>
-          </HStack>
-        </Flex>
-      </Box>
+      <MissingReferenceForm />
     </FeedbackLayout>
   );
 };
